pyproject.toml
[tool.poetry]
name = "rag-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
llama-index = "^0.10.26"
ipykernel = "^6.29.4"
llama-index-embeddings-huggingface = "^0.2.0"
pypdf2 = "^3.0.1"
load-dotenv = "^0.1.0"
ragas = "^0.1.6"
qdrant-client = "^1.8.2"
llama-index-vector-stores-qdrant = "^0.1.6"
llama-index-postprocessor-flag-embedding-reranker = "^0.1.2"
flagembedding = {git = "https://github.com/FlagOpen/FlagEmbedding.git"}
llama-index-postprocessor-voyageai-rerank = "^0.1.1"
arxiv = "^2.1.0"
feedparser = "6.0.10"
supabase = "^2.4.2"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
